flask-ai/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
requests==2.32.3
google-genai>=1.0.0
rapidfuzz>=3.6.0
